refactor: remove dictionary and implement `Name`
